refactor: refactor module names to avoid function/module name clashes (#114)
- Renamed:
  - `clone.py`       → `repository_clone.py`
  - `ingest.py`      → `repository_ingest.py`
  - `ingest_from_query.py` → `query_ingestion.py`
  - `parse_query.py` → `query_parser.py`
