+++ src/gitingest/__init__.py
@@ -1,8 +1,8 @@
 """ Gitingest: A package for ingesting data from git repositories. """
 
-from gitingest.clone import clone_repo
-from gitingest.ingest import ingest
-from gitingest.ingest_from_query import run_ingest_query
-from gitingest.parse_query import parse_query
+from gitingest.query_ingestion import run_ingest_query
+from gitingest.query_parser import parse_query
+from gitingest.repository_clone import clone_repo
+from gitingest.repository_ingest import ingest
 
 __all__ = ["run_ingest_query", "clone_repo", "parse_query", "ingest"]
