+++ src/routers/dynamic.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Form, Request
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
 
-from process_query import process_query
+from query_processor import process_query
 from server_utils import limiter
 
 router = APIRouter()
